\name{hd}
\docType{data}
\alias{hd}
\title{Hodgkin's Disease Data}
\description{
 Competing risk data set involving Hodgkin's disease (Table 1.6b, \emph{Competing Risks: A
 Practical Perspective}).
}
\references{
    Pintilie M., (2006) \emph{Competing Risks: A
      Practical Perspective.} West Sussex: John Wiley and Sons.
}
\keyword{datasets}
