\name{pbc}
\docType{data}
\alias{pbc}
\title{Primary Biliary Cirrhosis (PBC) Data}
\description{
  Data from the Mayo Clinic trial involving primary biliary cirrhosis (PBC).
  The first 312 cases participated in the randomized trial and contain
  largely complete data.  An additional 106 cases did not participate in
  the trial, but consented to have basic measurements recorded and to be
  followed for survival.  These have moderate amounts of missing data.
}
\references{
    Flemming T.R and Harrington D.P., (1991) \emph{Counting Processes
    and Survival Analysis.} New York: Wiley.
}
\keyword{datasets}
