\name{rsf.news}
\alias{rsf.news}
\title{Show the NEWS file}
\description{
    Show the NEWS file of the \pkg{randomSurvivalForest} package.
}
\usage{
    rsf.news(...)
}
\arguments{
    \item{...}{Further arguments passed to or from other methods.}
}
\value{
    None.
}
\author{
  Hemant Ishwaran \email{hemant.ishwaran@gmail.com}

  Udaya B. Kogalur \email{kogalurshear@gmail.com}
}
\keyword{documentation}
