\name{veteran}
\docType{data}
\alias{veteran}
\title{Veteran's Administration Lung Cancer Trial Data}
\description{
    Randomized trial of two treatment regimens for lung cancer.
    This is a standard survival analysis data set.
}
\references{
    Kalbfleisch J. and Prentice R, (1980) \emph{The Statistical
    Analysis of Failure Time Data.}  New York: Wiley.
}
\keyword{datasets}
