\name{breast}
\docType{data}
\alias{breast}
\title{German Breast Cancer Data}
\description{
  Breast cancer survival data.
}
\references{
  M. Schumacher, et al. (1994).  Randomized \eqn{2\times2} trial
  evaluating hormonal treatment and the duration of chemotherapy in
  node-positive breast cancer patients.  The German Breast Cancer Study
  Group, \emph{J. Clinical Oncology}, 12:2086-2093.
}
\keyword{datasets}
