\name{follic}
\docType{data}
\alias{follic}
\title{Follicular Cell Lymphoma Data}
\description{
  Competing risk data set involving follicular cell lymphoma (Table
  1.4b, \emph{Competing Risks: A Practical Perspective}).
}
\references{
    Pintilie M., (2006) \emph{Competing Risks: A
      Practical Perspective.} West Sussex: John Wiley and Sons.
}
\keyword{datasets}
