\name{rsf.news}
\alias{rsf.news}
\title{Show the NEWS file}
\description{
    Show the NEWS file of the \pkg{randomSurvivalForest} package.
}
\usage{
    rsf.news()
}
\value{
    None.
}
\author{
    Hemant Ishwaran \email{hemant.ishwaran@gmail.com} and
    Udaya B. Kogalur \email{ubk2101@columbia.edu}
}
\keyword{documentation}
